Refining the AutoTitle service
